requirements.txt
asgiref==3.8.1
Django~=5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
openai==1.51.0
pip==24.2
requests==2.32.3
sqlparse==0.5.1
transformers==4.45.1
